Refactor "Cluster" Spawners
The calculation for determining the X/Y offsets for the given scientist during the actual spawning of the cluster is now actually calculated by evenly distributing the angle by the number of scientists spawning, offset by a random amount at the point the cluster spawner starts.

Also use goto/loop/etc. instead of a 0-tic A_Jump with a 100% chance to jump.