More Cluster Spawner Refinements
The spawner now checks if the position is even valid when attempting to spawn the mob.  If it's not, destroy the actor and retry, up to 10 times.  If that threshold is ever met, just stop as it's possible there aren't any viable spawn positions (small alcove, tight space with many other mobs, etc.)
The spawner now also simply picks a random position around the spawner origin, rather than orienting them in a circle.  This feels much more natural, should those mobs spawn within view of the player.