Update FemaleSyringe.zsc
Adds a comedown attack for the Zombie Scientist Syringe monster. I tried to look for a specific spot to see if I could make sure you didn't get hit with a massive dose, but I got lost. 

Lines 59 was changed to allow for a larger range and 65 was added to make a new "drug" value